Featuretoggle nytt filter og få appen til å kompilere
