+++ src/søk/Søk.tsx
@@ -1,11 +1,11 @@
 import React, { FunctionComponent } from 'react';
 import Søkefelt from './søkefelt/Søkefelt';
-import HvorErAnnonsenPublisert from './HvorErAnnonsenPublisert';
 import { QueryParam, QueryParamValue } from './søkefelt/urlUtils';
 import Geografi from './geografi/Geografi';
 import { erIkkeProd } from '../utils/featureToggleUtils';
 import FylkerOgKommuner from './geografi/FylkerOgKommuner';
 import './Søk.less';
+import OmAnnonsen from './om-annonsen/OmAnnonsen';
 
 export type SøkProps = {
     oppdaterSøk: (queryParam: QueryParam, verdi: QueryParamValue) => void;
@@ -15,7 +15,7 @@ const Søk: FunctionComponent<SøkProps> = ({ oppdaterSøk }) => {
     return (
         <div className="søk">
             <Søkefelt oppdaterSøk={oppdaterSøk} />
-            <HvorErAnnonsenPublisert oppdaterSøk={oppdaterSøk} />
+            <OmAnnonsen oppdaterSøk={oppdaterSøk} />
             {erIkkeProd ? (
                 <FylkerOgKommuner oppdaterSøk={oppdaterSøk} />
             ) : (
